dma is broke
